requirements.txt
Django
channels
channels-redis
django-countries
psycopg2
djangorestframework
django-cors-headers
django-filebrowser
django-grappelli==2.13.3
django-image-cropping
easy-thumbnails
PythonVideoConverter==1.0.3
drf-yasg
django-rest-framework-social-oauth2
geoip2
django-grappelli==2.13.3
celery
django-celery-results
django-celery-beat
django-cleanup
django-rosetta 
django-modeltranslation==0.13.3
grappelli-modeltranslation==0.1.3